% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m6Am_call.r
\name{m6Am_call}
\alias{m6Am_call}
\title{m6Am_call}
\usage{
m6Am_call(IP_BAM, INPUT_BAM, KO_IP_BAM, KO_INPUT_BAM, GENE_ANNO_SAF = NA)
}
\arguments{
\item{IP_BAM:}{bam files for control ip sample}

\item{INPUT_BAM:}{bam files for control input sample}

\item{KO_IP_BAM:}{bam files for PCIF1 knockout ip sample}

\item{KO_INPUT_BAM:}{bam files for PCIF1 knockout input sample}

\item{GENE_ANNO_SAF:}{.saf file for the annotation of TSS regions}
}
\value{
a .csv file that lists fold change, log.p, log.fdr for reads in every TSS region across the different conditions
}
\description{
This function detects m6Am sites through the comparision of samples from m6A-exo-seq data
}
\details{
This is the description of m6Am_call function.

This function performs differential methylation analysis for samples under control and m6Am methyltransferase knockout conditions. Differentially methylated transcription start sites are defined as m6Am sites.
}
