% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtest.r
\name{rhtest}
\alias{rhtest}
\title{rhtest}
\usage{
rhtest(
  untreated_ip,
  untreated_input,
  treated_ip,
  treated_input,
  untreated_ip_total,
  untreated_input_total,
  treated_ip_total,
  treated_input_total,
  minimal_count_fdr = 10
)
}
\arguments{
\item{untreated_ip:}{bam files for control ip sample}

\item{untreated_input:}{bam files for control input sample}

\item{treated_ip:}{bam files for PCIF1 knockout ip sample}

\item{treated_input:}{bam files for PCIF1 knockout input sample}

\item{untreated_ip_total:}{depth of control ip sample}

\item{untreated_input_total:}{depth of control input sample}

\item{treated_ip_total:}{depth of PCIF1 knockout ip sample}

\item{treated_input_total:}{depth of PCIF1 knockout input sample}
}
\value{
a list object that lists fold change, log.p, log.fdr for reads in every TSS region across the different conditions
}
\description{
This function is the core funtion within m6Am_call
}
\details{
This is the description of rhtest function.

This function performs differential methylation analysis through hypergeometric test.
}
